# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: InfoShield API - Information Warfare Learning App
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quizzes:
    get:
      operationId: listQuizzes
      tags: [quizzes]
      summary: List all quizzes
      parameters:
        - name: moduleId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of quizzes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quiz"
  /quizzes/{quizId}:
    get:
      operationId: getQuiz
      tags: [quizzes]
      summary: Get quiz with questions
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizWithQuestions"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /quizzes/{quizId}/submit:
    post:
      operationId: submitQuiz
      tags: [quizzes]
      summary: Submit quiz answers and get results
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSubmission"
      responses:
        "200":
          description: Quiz result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Quiz:
      type: object
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        description:
          type: string
        questionCount:
          type: integer
        passingScore:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - moduleId
        - title
        - description
        - questionCount
        - passingScore
        - createdAt
    QuizQuestion:
      type: object
      properties:
        id:
          type: integer
        quizId:
          type: integer
        question:
          type: string
        options:
          type: array
          items:
            type: string
        hint:
          type: string
        orderIndex:
          type: integer
      required:
        - id
        - quizId
        - question
        - options
        - orderIndex
    QuizWithQuestions:
      type: object
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        description:
          type: string
        passingScore:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - moduleId
        - title
        - description
        - passingScore
        - questions
        - createdAt
    QuizSubmission:
      type: object
      properties:
        userId:
          type: string
        answers:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: integer
              selectedOption:
                type: integer
            required:
              - questionId
              - selectedOption
      required:
        - userId
        - answers
    QuizResult:
      type: object
      properties:
        quizId:
          type: integer
        userId:
          type: string
        score:
          type: integer
        totalQuestions:
          type: integer
        passed:
          type: boolean
        xpEarned:
          type: integer
        badgesEarned:
          type: array
          items:
            $ref: "#/components/schemas/Badge"
        answerFeedback:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: integer
              correct:
                type: boolean
              correctOption:
                type: integer
              explanation:
                type: string
            required:
              - questionId
              - correct
              - correctOption
      required:
        - quizId
        - userId
        - score
        - totalQuestions
        - passed
        - xpEarned
        - badgesEarned
        - answerFeedback
    Badge:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        rarity:
          type: string
          enum: [common, uncommon, rare, legendary]
        xpValue:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - icon
        - rarity
        - xpValue
        - createdAt
...