# lib/api-zod/src/generated/types/quiz.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * InfoShield API - Information Warfare Learning App
 * OpenAPI spec version: 0.1.0
 */

export interface Quiz {
  id: number;
  moduleId: number;
  title: string;
  description: string;
  questionCount: number;
  passingScore: number;
  createdAt: Date;
}