# lib/api-zod/src/generated/types/quizWithQuestions.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * InfoShield API - Information Warfare Learning App
 * OpenAPI spec version: 0.1.0
 */
import type { QuizQuestion } from "./quizQuestion";

export interface QuizWithQuestions {
  id: number;
  moduleId: number;
  title: string;
  description: string;
  passingScore: number;
  questions: QuizQuestion[];
  createdAt: Date;
}